Improved naming and added comments
