+++ twitchVotingOverlay/src/chaosOverlayClient/iMessage.ts
@@ -1,0 +1,8 @@
+import { IChaosOverlayVoteOption } from './iVoteOption';
+
+export interface IChaosOverlayClientMessage {
+	request: 'CREATE' | 'UPDATE';
+	totalVotes: number;
+	votingMode: 'MAJORITY' | 'PERCENTAGE';
+	voteOptions: IChaosOverlayVoteOption[];
+}
