+++ twitchVotingOverlay/src/chaosOverlayClient/iVoteOption.ts
@@ -1,4 +1,4 @@
-export interface IVoteOption {
+export interface IChaosOverlayVoteOption {
 	/**
 	 * Name for the option
 	 */
