+++ twitchVotingOverlay/src/chaosOverlayClient/tEvent.ts
@@ -1,0 +1,3 @@
+import { IChaosOverlayClientMessage } from './iMessage';
+
+export type TChaosOverlayClientEvent = (event: IChaosOverlayClientMessage) => void;
