--- twitchVotingOverlay/src/chaosWSClient/iMessage.ts
@@ -1,8 +1,0 @@
-import { IVoteOption } from './iVoteOption';
-
-export interface IMessage {
-	request: 'CREATE' | 'UPDATE';
-	totalVotes: number;
-	votingMode: 'MAJORITY' | 'PERCENTAGE';
-	voteOptions: IVoteOption[];
-}
