--- twitchVotingOverlay/src/chaosWSClient/tMessageListener.ts
@@ -1,3 +1,0 @@
-import { IMessage } from './iMessage';
-
-export type TMessageListener = (event: IMessage) => void;
